Increased allowed rupture top depth param depth to 100km

+++ src/org/opensha/sha/imr/param/EqkRuptureParams/RupTopDepthParam.java
@@ -32,7 +32,7 @@ public class RupTopDepthParam extends WarningDoubleParameter {
 	public final static String UNITS = "km";
 	public final static String INFO = "The depth to the shallowest point on the earthquake rupture surface";
 	protected final static Double MIN = new Double(0);
-	protected final static Double MAX = new Double(30);
+	protected final static Double MAX = new Double(100);
 	// warning values are set in subclasses
